requirements.txt
markdown
google-generativeai
python-frontmatter
jinja2
requests
pyyaml
